pyproject.toml
[tool.poetry]
name = "skallelego"
version = "0.1.0"
description = ""
authors = ["maxibor <maxime_borry@eva.mpg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-allel = "^1.3.13"
numpy = "^1.26.4"
vcfpy = "^0.13.8"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
